+++ exo-go/src/runner/runner.go
@@ -1,4 +1,4 @@
-package run
+package runner
 
 import (
 	"fmt"
