+++ exo-go/src/runner/service_restarter.go
@@ -1,4 +1,4 @@
-package run
+package runner
 
 import (
 	"fmt"
